requirements.txt
Django>=3.1.6
django-oauth-toolkit~=1.4.0
djangorestframework~=3.12.2
django-cors-middleware
Pillow
django-braces
requests
psycopg2